fix: addressing pr comments
