+++ app/components/Views/confirmations/components/rows/bridge-fee-row/bridge-fee-row.tsx
@@ -26,7 +26,9 @@ import { RowAlertKey } from '../../UI/info-row/alert-row/constants';
 import { useAlerts } from '../../../context/alert-system-context';
 import useFiatFormatter from '../../../../../UI/SimulationDetails/FiatDisplay/useFiatFormatter';
 import { ConfirmationRowComponentIDs } from '../../../ConfirmationView.testIds';
-import { NetworkFeeRow } from '../network-fee-row';
+import { IconColor } from '../../../../../../component-library/components/Icons/Icon';
+
+const NETWORK_FEE_ONLY_TYPES = [TransactionType.musdConversion];
 
 export function BridgeFeeRow() {
   const transactionMetadata = useTransactionMetadataOrThrow();
@@ -36,6 +38,15 @@ export function BridgeFeeRow() {
   const totals = useTransactionPayTotals();
   const { fieldAlerts } = useAlerts();
   const hasAlert = fieldAlerts.some((a) => a.field === RowAlertKey.PayWithFee);
+
+  const networkFeeUsd = useMemo(() => {
+    const sourceNetworkUsd = totals?.fees?.sourceNetwork?.estimate?.usd;
+    const targetNetworkUsd = totals?.fees?.targetNetwork?.usd;
+
+    if (sourceNetworkUsd == null || targetNetworkUsd == null) return '';
+
+    return formatFiat(new BigNumber(sourceNetworkUsd).plus(targetNetworkUsd));
+  }, [totals, formatFiat]);
 
   const feeTotalUsd = useMemo(() => {
     if (!totals?.fees) return '';
@@ -52,10 +63,30 @@ export function BridgeFeeRow() {
     [formatFiat],
   );
 
-  if (
-    hasTransactionType(transactionMetadata, [TransactionType.musdConversion])
-  ) {
-    return <NetworkFeeRow />;
+  if (hasTransactionType(transactionMetadata, NETWORK_FEE_ONLY_TYPES)) {
+    if (isLoading) {
+      return <InfoRowSkeleton testId="bridge-fee-row-skeleton" />;
+    }
+
+    return (
+      <AlertRow
+        testID="bridge-fee-row-skeleton"
+        label={strings('confirm.label.network_fee')}
+        alertField={RowAlertKey.PayWithFee}
+        tooltipTitle={strings('confirm.label.network_fee')}
+        tooltip={strings('confirm.tooltip.network_fee')}
+        tooltipColor={IconColor.Alternative}
+        rowVariant={InfoRowVariant.Small}
+      >
+        <Text
+          variant={TextVariant.BodyMD}
+          color={hasAlert ? TextColor.Error : TextColor.Alternative}
+          testID={ConfirmationRowComponentIDs.NETWORK_FEE}
+        >
+          {networkFeeUsd}
+        </Text>
+      </AlertRow>
+    );
   }
 
   if (isLoading) {
