--- app/components/Views/confirmations/components/rows/network-fee-row/index.ts
@@ -1,1 +1,0 @@
-export * from './network-fee-row';
